requirements.txt
asgiref==3.8.1
certifi @ file:///C:/ci/certifi_1655889035820/work/certifi
Django==4.2.15
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
importlib_metadata==8.2.0
Markdown==3.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
wincertstore==0.2
zipp==3.20.0
